made changes to the Code bases types files also waiting on zack/sean for tommorrow

+++ shift-scan/src/actions/inboxSentActions.tsx
@@ -1,9 +1,9 @@
-"use server"
+"use server";
+
+import prisma from "@/lib/prisma";
 import { FormStatus, PrismaClient } from "@prisma/client";
-import { request } from "http";
-import { revalidatePath } from "next/cache";
+import { revalidatePath } from "next/cache"
 import { redirect } from "next/navigation";
-const prisma = new PrismaClient();
 
 export async function getUserSentContent(user_id: string | undefined) {
     if (!user_id) return
@@ -45,22 +45,58 @@ export async function createLeaveRequest(formData: FormData) {
 }
 
 export async function EditLeaveRequest(formData: FormData) {
+    try {
+        console.log(formData)
+        const user_id = formData.get("user_id") as string;
+        const requestType = formData.get("requestType") as string;
+        const status = formData.get("status") as string;
+        
+        if (status) {
+            const result = await prisma.timeoffRequestForm.update({
+            where: { id: Number(formData.get("id") as string) },
+            data: {
+            date: new Date(formData.get("date") as string),
+            requestedStartDate: new Date(formData.get("startDate") as string),
+            requestedEndDate: new Date(formData.get("endDate") as string),
+            requestType: requestType,
+            comments: formData.get("description") as string,
+            mangerComments : null,
+            status: status as FormStatus,
+            employee_id: user_id,
+            }
+            })
+            console.log(result)
+            revalidatePath("/hamburger/inbox");
+        }
+    } catch (error) {
+        console.log(error)
+    }
 
 }
 
-export async function DeleteLeaveRequest(formData: FormData) {
+export async function DeleteLeaveRequest(id : string, user_id : string | undefined) {
     try {
-        console.log(formData);
+        const deleteId = Number(id)
 
-        const id = Number(formData.get('id') as string);
+    const userCheck = await prisma.timeoffRequestForm.findMany({
+        where: {
+            id: deleteId,
+            employee_id: user_id
+        }
+    })
 
+    if (userCheck.length > 0) {
         await prisma.timeoffRequestForm.delete({
-            where: { id: id },
+            where: { id: deleteId },
         });
-
         revalidatePath("/hamburger/inbox");
         redirect("/hamburger/inbox");
-    } catch (error) {
-        console.error("Failed to delete request:", error);
     }
+    else {
+        throw new Error("You do not have permission to delete this request");
+    }
+
+} catch (error) {
+    console.error("Failed to delete request:", error);
+}
 }
